refactor: cleanup comments
