+++ processor/utils.ts
@@ -13,11 +13,18 @@ import { CONTINUE, EXIT, visit } from 'unist-util-visit';
 import mdast from '../lib/mdast';
 
 /**
- * Evaluate a JSX attribute expression source as a plain JavaScript literal.
- * Wrapping in parens lets object literals (`{color: 'red'}`) parse as expressions.
- * Runs with no scope, so only self-contained literals resolve.
+ * Evaluate a JavaScript expression source and return its value.
+ *
+ * Wrapping in parens lets object literals (`{color: 'red'}`) parse as
+ * expressions. Runs with no scope, so only self-contained literals resolve;
+ * anything that fails to parse or throws at runtime falls back to the raw source.
+ *
+ * > ☢️ **Danger**: this `eval`s JavaScript. Only call when safeMode is off —
+ * > safeMode's contract is that expression syntax is never evaluated, and the
+ * > pipeline guards against reaching this path by keeping attribute expressions
+ * > as literal strings and skipping the expression transformer entirely.
  */
-export function evaluateAttributeExpression(source: string): unknown {
+export function evaluate(source: string): unknown {
   try {
     // eslint-disable-next-line no-new-func
     return new Function(`return (${source})`)();
@@ -90,7 +97,7 @@ export const getAttrs = <T>(jsx: MdxJsxFlowElement | MdxJsxTextElement): T =>
       } else if (typeof attr.value === 'string') {
         memo[attr.name] = attr.value;
       } else if (attr.value?.value !== undefined) {
-        memo[attr.name] = evaluateAttributeExpression(attr.value.value);
+        memo[attr.name] = evaluate(attr.value.value);
       }
     }
